+++ CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
  - Added port configuration features to devices and ports
 
+## [6.1.0] - 2019-04-19
+### Added
+ - Cresnet Occupancy Sensor Adapters - GLS-ODT-C-CN, GLS-OIR-C-CN, and GLS-OIRLCL-C-CN
+
 ## [6.0.2] - 2019-01-16
 ### Changed
  - Less obnoxious RelayPortAdapter error when failing to get the internal port